Moved public classes to exported packages + minor fixes.
