+++ core/roboconf-messaging-http/pom.xml
@@ -22,8 +22,7 @@
    distributed under the License is distributed on an "AS IS" BASIS,
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
-   limitations under the License.age governing permissions and limitations
-  ~ under the License.
+   limitations under the License.
   -->
 
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
@@ -57,13 +56,6 @@
 		</dependency>
 
 		<dependency>
-			<groupId>net.roboconf</groupId>
-			<artifactId>roboconf-dm-rest-commons</artifactId>
-			<version>${project.version}</version>
-			<scope>provided</scope>
-		</dependency>
-
-		<dependency>
 			<groupId>org.apache.felix</groupId>
 			<artifactId>org.apache.felix.ipojo.annotations</artifactId>
 			<version>${ipojo.version}</version>
@@ -83,29 +75,45 @@
 			<version>4.3.1</version>
 			<scope>provided</scope>
 		</dependency>
-
+		
 		<dependency>
 			<groupId>javax.websocket</groupId>
 			<artifactId>javax.websocket-api</artifactId>
 			<version>1.0</version>
+			<scope>provided</scope>
 		</dependency>
-		<!-- To run javax.websocket in embedded server -->
+
+	<!-- To run javax.websocket in embedded server -->
 		<dependency>
 			<groupId>org.eclipse.jetty.websocket</groupId>
 			<artifactId>javax-websocket-server-impl</artifactId>
-			<version>9.2.7.v20150116</version>
+			<version>${websocket.version}</version>
 		</dependency>
 		<!-- To run javax.websocket client -->
 		<dependency>
 			<groupId>org.eclipse.jetty.websocket</groupId>
 			<artifactId>javax-websocket-client-impl</artifactId>
-			<version>9.2.7.v20150116</version>
+			<version>${websocket.version}</version>
 		</dependency>
 
 		<dependency>
-			<groupId>com.sun.jersey</groupId>
-			<artifactId>jersey-client</artifactId>
-			<version>${jersey.stack.version}</version>
+			<groupId>org.eclipse.jetty.websocket</groupId>
+			<artifactId>websocket-api</artifactId>
+			<version>${websocket.version}</version>
+			<scope>provided</scope>
+		</dependency>
+		
+		<dependency>
+			<groupId>org.eclipse.jetty.websocket</groupId>
+			<artifactId>websocket-servlet</artifactId>
+			<version>${websocket.version}</version>
+			<scope>provided</scope>
+		</dependency>
+		
+		<dependency>
+			<groupId>javax.servlet</groupId>
+			<artifactId>servlet-api</artifactId>
+			<version>2.5</version>
 			<scope>provided</scope>
 		</dependency>
 
@@ -148,15 +156,18 @@
 					<instructions>
 						<Import-Package>
 							net.roboconf.*;version="${project.version}",
+							javax.annotation,
 							javax.servlet,
-							org.osgi.service.http; version="[1.2,1.3)",
-							org.osgi.framework; version="[1.6,2.0)",
+							javax.servlet.http,
+							*
+							<!--  org.osgi.service.http; version="[1.2,1.3)",
+							org.osgi.framework; version="[1.6,2.0)", -->
 						</Import-Package>
 						<Export-Package>
-							net.roboconf.messaging-http; version="${project.version}"
+							net.roboconf.messaging.http; version="${project.version}"
 						</Export-Package>
 						<Private-Package>
-							net.roboconf.messaging-http.internal
+							net.roboconf.messaging.http.internal
 						</Private-Package>
 					</instructions>
 				</configuration>
