+++ karaf/roboconf-karaf-dist-dm/karaf-resources/net.roboconf.messaging.http.cfg
@@ -21,5 +21,5 @@
 
 # The IP address (or hostname) and port of http server.
 net.roboconf.messaging.http.server.ip = localhost
-net.roboconf.messaging.http.server.port = 8080
+net.roboconf.messaging.http.server.port = 8181
